# TypeInfo.GetAncestors() -> TypeInfo.GetAncestor(int index)