--- a/docs2/site/docs/migrations/migration4.md
+++ b/docs2/site/docs/migrations/migration4.md
@@ -26,12 +26,6 @@
 * `IHaveDefaultValue.Type` has been moved to `IProvideResolvedType.Type`
 * `Connection<TNode, TEdge>.TotalCount` has been changed from an `int` to an `int?`. This allows for returning `null` indicating that the total count is unknown.
 * By default fields returned by introspection query are no longer sorted by their names. `LegacyV3SchemaComparer` can be used to switch to the old behavior.
-* `INode.IsEqualTo` and related methods have been removed.
-* `NameNode` has been changed to a readonly struct.
-* `DebugNodeVisitor` class has been removed.
-* Most methods and classes within `OverlappingFieldsCanBeMerged` are now private.
-* `EnumerableExtensions.Apply` for dictionaries has been removed.
-* `ObjectExtensions.GetInterface` has been removed along with two overloads of `GetPropertyValue`.
 
 ```c#
     /// <summary>
@@ -52,5 +46,12 @@
     }
 ```
 
+* `INode.IsEqualTo` and related methods have been removed.
+* `NameNode` has been changed to a readonly struct.
+* `DebugNodeVisitor` class has been removed.
+* Most methods and classes within `OverlappingFieldsCanBeMerged` are now private.
+* `EnumerableExtensions.Apply` for dictionaries has been removed.
+* `ObjectExtensions.GetInterface` has been removed along with two overloads of `GetPropertyValue`.
 * Subscriptions implementation (`SubscriptionExecutionStrategy`) has been moved into `GraphQL.SystemReactive` project and default document executer throws `NotSupportedException`.
 * `ISubscriptionExecuter` has been removed.
+* `TypeInfo.GetAncestors()` has been changed to `TypeInfo.GetAncestor(int index)`