--- a/src/GraphQL.Harness/Properties/launchSettings.json
+++ b/src/GraphQL.Harness/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "Playground": {
       "commandName": "Project",
-      "launchBrowser": true,
+      "launchBrowser": false,
       "launchUrl": "ui/playground",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"