fixes #2152

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.16299.726 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i7 CPU 920 2.67GHz (Nehalem), 1 CPU, 8 logical and 4 physical cores
Frequency=2607421 Hz, Resolution=383.5207 ns, Timer=TSC
.NET Core SDK=5.0.101
  [Host]     : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT
  DefaultJob : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT
```

```
before

|        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| Introspection | 553.22 μs | 1.922 μs | 1.798 μs | 7.8125 |     - |     - |  34.13 KB |
|          Hero |  46.22 μs | 0.053 μs | 0.047 μs | 3.7231 |     - |     - |   15.3 KB |

after

|        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| Introspection | 573.30 μs | 0.944 μs | 0.837 μs | 7.8125 |     - |     - |  33.44 KB |
|          Hero |  45.16 μs | 0.036 μs | 0.030 μs | 3.7231 |     - |     - |   15.3 KB |

after remove try/finally

|        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| Introspection | 553.22 μs | 0.710 μs | 0.664 μs | 7.8125 |     - |     - |  33.44 KB |
|          Hero |  46.38 μs | 0.099 μs | 0.087 μs | 3.7231 |     - |     - |   15.3 KB |
```

You can try it yourself. Execution time is confusing again.