@Shane32 2021-01-08
I will try it later.  If it's working slower than LINQ, I'd be quite surprised.  I think you have the fastest implementation right now.  But it is possible that Stack.ToArray is faster than using an enumerator.  And if it is, it's probably due to less jumps and conditionals within the compiled assembly language instructions, so it doesn't have to empty the CPU pipeline as often.  I'll try it.