@Shane32 2021-01-08
I ran benchmarks:
- `tune` branch
- the develop branch as it appears in the `tune` branch history
- the develop branch as it appears in the `tune` branch history, along with the changes to GetAncestors that were in #2139 (that are now in the current develop branch)

```
tune

|        Method |      Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|-------:|------:|----------:|
| Introspection | 237.20 us | 0.540 us | 0.505 us | 5.3711 |      - |     - |  33.44 KB |
|     Fragments | 102.13 us | 0.243 us | 0.215 us | 3.6621 | 0.1221 |     - |   22.8 KB |
|          Hero |  19.98 us | 0.242 us | 0.227 us | 2.4719 | 0.0610 |     - |   15.3 KB |


develop (pre-validation_3)

|        Method |      Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|-------:|------:|----------:|
| Introspection | 239.25 us | 0.742 us | 0.620 us | 5.3711 |      - |     - |  34.13 KB |
|          Hero |  19.92 us | 0.205 us | 0.172 us | 2.4719 | 0.0610 |     - |   15.3 KB |


develop (pre-validation_3 with revised getAncestors from validation_3)


|        Method |      Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|-------:|------:|----------:|
| Introspection | 232.62 us | 0.698 us | 0.653 us | 5.3711 |      - |     - |   33.7 KB |
|          Hero |  19.54 us | 0.070 us | 0.066 us | 2.4719 | 0.0610 |     - |   15.3 KB |
```

I'm fine with whatever changes you want to make to this method, btw. You decide.  If you want me to run more benchmarks, let me know.