@sungam3r 2021-01-08
I think we should accept these changes. The more complex (nested) query, the greater the gain due to the fact that arrays are not copied. The most recent elements of the stack are always used.