@Shane32 2021-01-08
I also though of this code:
```cs
        public void GetAncestors(out INode parent, out INode grandparent)
        {
            var ancestorArray = _ancestorStack.ToArray();
            parent = ancestorArray[1];
            grandparent = ancestorArray[2];
        }
```
But this tested slower:
```
|        Method |      Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|-------:|------:|----------:|
| Introspection | 238.93 us | 1.084 us | 1.014 us | 5.3711 |      - |     - |  33.58 KB |
|          Hero |  19.18 us | 0.098 us | 0.092 us | 2.4719 | 0.0610 |     - |   15.3 KB |
```
Which makes no sense.  Maybe it's that ref/out parameters are slow. I dunno