@sungam3r 2021-01-08
I believe that a slight drop in performance of the most primitive queries can be allowed in favor of general normalization of work on an arbitrary query size.