@Shane32 2021-01-08
I'd also like to comment that in the one place it's used, you are allocating an enumeration object twice just to read the first and second entries.