@sungam3r 2021-01-08
Stack enumerator is struct.