@Shane32 2021-01-08
> Did you test my changes with try/finally removed ?

Yes I just ran them - it had this history at the time I ran it:

![image](https://user-images.githubusercontent.com/6377684/103994477-102f1580-5165-11eb-96dd-026059157572.png)