@Shane32 2021-01-08
> Stack enumerator is struct.

Oh well that's cool! 👍