@Shane32 2021-01-08
I tested this also:
```cs
        public (INode parent, INode grandparent) GetAncestors()
        {
            var ancestorArray = _ancestorStack.ToArray();
            return (ancestorArray[1], ancestorArray[2]);
        }
```

Still repeatably tests slower than this code:
```cs
        public INode[] GetAncestors()
        {
            var ancestorArray = _ancestorStack.ToArray();
            var c = ancestorArray.Length;
            var ret = new INode[--c];
            for (int i = c; i > 0; --i)
            {
                ret[c - i] = ancestorArray[i];
            }
            return ret;
            // aka: return _ancestorStack.Skip(1).Reverse().ToArray();
        }
```

Weird.

Yeah merge whatever you want 👍