@Shane32 2021-01-08
```
tune

|        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| Introspection | 180.17 us | 0.371 us | 0.347 us | 3.1738 |     - |     - |  19.59 KB |
|     Fragments |  80.66 us | 0.378 us | 0.354 us | 1.9531 |     - |     - |  12.31 KB |
|          Hero |  12.59 us | 0.028 us | 0.025 us | 0.7935 |     - |     - |   4.95 KB |

develop

|        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| Introspection | 182.71 us | 0.312 us | 0.292 us | 3.1738 |     - |     - |  19.85 KB |
|          Hero |  12.55 us | 0.036 us | 0.034 us | 0.7935 |     - |     - |   4.95 KB |
```