@Shane32 2021-01-08
only 1% of validation, but hey, every 1% counts! and about 1.5% memory reduction (of validation, for large queries).